only show snapshot/volume names when stdout is a pipe
I often use this script as a handy way to list all snapshots without
having to remember the zfs(1) syntax, so just provide a plain list of
all names when not writing to the terminal (Or more exactly: When
writing to a pipe.  Writing to a file still produces full output.).